// apps/amplify-be/src/graphql/API.ts
export type CreateUserInput = {
  username: string,
  email: string,
  userProfileUsername?: string | null,
};

export type ModelUserConditionInput = {
  email?: ModelStringInput | null,
  and?: Array< ModelUserConditionInput | null > | null,
  or?: Array< ModelUserConditionInput | null > | null,
  not?: ModelUserConditionInput | null,
  userProfileUsername?: ModelStringInput | null,
};
// ...
export type User = {
  __typename: "User",
  username: string,
  email: string,
  profile?: Profile | null,
  notifications?: ModelNotificationConnection | null,
  createdAt: string,
  updatedAt: string,
  userProfileUsername?: string | null,
};

export type Profile = {
  __typename: "Profile",
  username: string,
  bio?: string | null,
  user?: User | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelNotificationConnection = {
  __typename: "ModelNotificationConnection",
  items:  Array<Notification | null >,
  nextToken?: string | null,
};

export type Notification = {
  __typename: "Notification",
  id: string,
  username: string,
  title: string,
  message: string,
  read: boolean,
  user?: User | null,
  createdAt: string,
  updatedAt: string,
  userNotificationsUsername?: string | null,
};

export type UpdateUserInput = {
  username: string,
  email?: string | null,
  userProfileUsername?: string | null,
};

export type DeleteUserInput = {
  username: string,
};

export type CreateProfileInput = {
  username: string,
  bio?: string | null,
};

export type ModelProfileConditionInput = {
  bio?: ModelStringInput | null,
  and?: Array< ModelProfileConditionInput | null > | null,
  or?: Array< ModelProfileConditionInput | null > | null,
  not?: ModelProfileConditionInput | null,
};

export type UpdateProfileInput = {
  username: string,
  bio?: string | null,
};

export type DeleteProfileInput = {
  username: string,
};

export type CreateNotificationInput = {
  id?: string | null,
  username: string,
  title: string,
  message: string,
  read: boolean,
  userNotificationsUsername?: string | null,
};
// ...
export type UpdateNotificationInput = {
  id: string,
  username?: string | null,
  title?: string | null,
  message?: string | null,
  read?: boolean | null,
  userNotificationsUsername?: string | null,
};
// ...
export type CreateUserMutationVariables = {
  input: CreateUserInput,
  condition?: ModelUserConditionInput | null,
};

export type CreateUserMutation = {
  createUser?:  {
    __typename: "User",
    username: string,
    email: string,
    profile?:  {
      __typename: "Profile",
      username: string,
      bio?: string | null,
      user?:  {
        __typename: "User",
        username: string,
        email: string,
        createdAt: string,
        updatedAt: string,
        userProfileUsername?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    notifications?:  {
      __typename: "ModelNotificationConnection",
      items:  Array< {
        __typename: "Notification",
        id: string,
        username: string,
        title: string,
        message: string,
        read: boolean,
        createdAt: string,
        updatedAt: string,
        userNotificationsUsername?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    userProfileUsername?: string | null,
  } | null,
};

export type UpdateUserMutationVariables = {
  input: UpdateUserInput,
  condition?: ModelUserConditionInput | null,
};

export type UpdateUserMutation = {
  updateUser?:  {
    __typename: "User",
    username: string,
    email: string,
    profile?:  {
      __typename: "Profile",
      username: string,
      bio?: string | null,
      user?:  {
        __typename: "User",
        username: string,
        email: string,
        createdAt: string,
        updatedAt: string,
        userProfileUsername?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    notifications?:  {
      __typename: "ModelNotificationConnection",
      items:  Array< {
        __typename: "Notification",
        id: string,
        username: string,
        title: string,
        message: string,
        read: boolean,
        createdAt: string,
        updatedAt: string,
        userNotificationsUsername?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    userProfileUsername?: string | null,
  } | null,
};

export type DeleteUserMutationVariables = {
  input: DeleteUserInput,
  condition?: ModelUserConditionInput | null,
};

export type DeleteUserMutation = {
  deleteUser?:  {
    __typename: "User",
    username: string,
    email: string,
    profile?:  {
      __typename: "Profile",
      username: string,
      bio?: string | null,
      user?:  {
        __typename: "User",
        username: string,
        email: string,
        createdAt: string,
        updatedAt: string,
        userProfileUsername?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    notifications?:  {
      __typename: "ModelNotificationConnection",
      items:  Array< {
        __typename: "Notification",
        id: string,
        username: string,
        title: string,
        message: string,
        read: boolean,
        createdAt: string,
        updatedAt: string,
        userNotificationsUsername?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    userProfileUsername?: string | null,
  } | null,
};

export type CreateProfileMutationVariables = {
  input: CreateProfileInput,
  condition?: ModelProfileConditionInput | null,
};

export type CreateProfileMutation = {
  createProfile?:  {
    __typename: "Profile",
    username: string,
    bio?: string | null,
    user?:  {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type UpdateProfileMutationVariables = {
  input: UpdateProfileInput,
  condition?: ModelProfileConditionInput | null,
};

export type UpdateProfileMutation = {
  updateProfile?:  {
    __typename: "Profile",
    username: string,
    bio?: string | null,
    user?:  {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type DeleteProfileMutationVariables = {
  input: DeleteProfileInput,
  condition?: ModelProfileConditionInput | null,
};

export type DeleteProfileMutation = {
  deleteProfile?:  {
    __typename: "Profile",
    username: string,
    bio?: string | null,
    user?:  {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type CreateNotificationMutationVariables = {
  input: CreateNotificationInput,
  condition?: ModelNotificationConditionInput | null,
};

export type CreateNotificationMutation = {
  createNotification?:  {
    __typename: "Notification",
    id: string,
    username: string,
    title: string,
    message: string,
    read: boolean,
    user?:  {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    userNotificationsUsername?: string | null,
  } | null,
};

export type UpdateNotificationMutationVariables = {
  input: UpdateNotificationInput,
  condition?: ModelNotificationConditionInput | null,
};

export type UpdateNotificationMutation = {
  updateNotification?:  {
    __typename: "Notification",
    id: string,
    username: string,
    title: string,
    message: string,
    read: boolean,
    user?:  {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    userNotificationsUsername?: string | null,
  } | null,
};

export type DeleteNotificationMutationVariables = {
  input: DeleteNotificationInput,
  condition?: ModelNotificationConditionInput | null,
};

export type DeleteNotificationMutation = {
  deleteNotification?:  {
    __typename: "Notification",
    id: string,
    username: string,
    title: string,
    message: string,
    read: boolean,
    user?:  {
      __typename: "User",
      username: string,
      email: string,
      profile?:  {
        __typename: "Profile",
        username: string,
        bio?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      notifications?:  {
        __typename: "ModelNotificationConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      userProfileUsername?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    userNotificationsUsername?: string | null,
  } | null,
};

export type CreateTransferMutationVariables = {
  input: CreateTransferInput,
  condition?: ModelTransferConditionInput | null,
};

export type CreateTransferMutation = {
  createTransfer?:  {
    __typename: "Transfer",
    id: string,
    from: string,
    to: string,
    amount: number,
    status: TransferStatus,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type UpdateTransferMutationVariables = {
  input: UpdateTransferInput,
  condition?: ModelTransferConditionInput | null,
};

export type UpdateTransferMutation = {
  updateTransfer?:  {
    __typename: "Transfer",
    id: string,
    from: string,
    to: string,
    amount: number,
    status: TransferStatus,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type DeleteTransferMutationVariables = {
  input: DeleteTransferInput,
  condition?: ModelTransferConditionInput | null,
};

export type DeleteTransferMutation = {
  deleteTransfer?:  {
    __typename: "Transfer",
    id: string,
    from: string,
    to: string,
    amount: number,
    status: TransferStatus,
    createdAt: string,
    updatedAt: string,
  } | null,
};